+++ app/src/main/java/com/zhiyicx/thinksnsplus/modules/chat/ChatPresenter.java
@@ -96,7 +96,7 @@ public class ChatPresenter extends BasePresenter<ChatContract.Repository, ChatCo
     @Override
     public void sendTextMessage(String text, int cid) {
         Message message = ChatClient.getInstance(mContext).sendTextMsg(text, cid, "");// usid 暂不使用
-        message.setUid(AppApplication.getmCurrentLoginAuth() != null ? AppApplication.getmCurrentLoginAuth().getUser_id() : 0);// 更新
+        message.setUid(AppApplication.getmCurrentLoginAuth() != null ? (int) AppApplication.getmCurrentLoginAuth().getUser_id() : 0);// 更新
         if (!ZBIMClient.getInstance().isLogin()) { // IM 没有连接成功
             message.setSend_status(MessageStatus.SEND_FAIL);
         }
@@ -131,7 +131,7 @@ public class ChatPresenter extends BasePresenter<ChatContract.Repository, ChatCo
                 .subscribe(new BaseSubscribe<Conversation>() {
                     @Override
                     protected void onSuccess(Conversation data) {
-                        data.setIm_uid(AppApplication.getmCurrentLoginAuth().getUser_id());
+                        data.setIm_uid((int) AppApplication.getmCurrentLoginAuth().getUser_id());
                         data.setUsids(uids);
                         data.setPair(pair);
                         mRepository.insertOrUpdateConversation(data);
@@ -195,7 +195,7 @@ public class ChatPresenter extends BasePresenter<ChatContract.Repository, ChatCo
         ChatItemBean chatItemBean = new ChatItemBean();
         chatItemBean.setLastMessage(message);
         if (message.getUid() == 0) {// 如果没有 uid, 则表明是当前用户发的消息
-            message.setUid(AppApplication.getmCurrentLoginAuth() != null ? AppApplication.getmCurrentLoginAuth().getUser_id() : 0);
+            message.setUid(AppApplication.getmCurrentLoginAuth() != null ? (int) AppApplication.getmCurrentLoginAuth().getUser_id() : 0);
         }
         UserInfoBean userInfoBean = mUserInfoBeanSparseArray.get(message.getUid());
         if (userInfoBean == null) {
